// src/interfaces/common.ts
import { IGenericErrorMessage } from "./error";


export type IGenericResponse<T> = {
  meta: {
    page: number;
    limit: number | undefined;
    total: number;
  };
  data: T;
};

export type IGenericErrorResponse = {
  statusCode: number;
  message: string;
  errorMessages: IGenericErrorMessage[];
};